Removed a separate header related to json db
